Migrate Realtime init tests
